hotfix : 비동기 순환참조문제 해결

+++ FE/danbi/src/Util/apis/api.js
@@ -11,54 +11,84 @@ export const setTokenExpireTime = (payload) => {
   token.setAccessTokenExpireTime(payload);
 }
 
-export const reissueAccessToken = ()=>{
+export const reissueAccessToken = async ()=>{
   const refreshToken = localStorage.getItem("refreshToken");
   
-  if(!token.check() || !refreshToken || refreshToken===""){
-    return -1;
+  if(!refreshToken || refreshToken===""){
+    return null;
   }
 
-  axios({
-    method:"post",
-    url:"/api/v1/access-token/issue",
-    headers:{"Authorization" : `Bearer ${refreshToken}`}
-  }).then(({data})=>{
+  try{
+    const {data} = await axios.post("/api/v1/access-token/issue",{
+      headers:{"Authorization" : `Bearer ${refreshToken}`}
+    });
     token.setAccessToken(data.accessToken);
     token.setAccessTokenExpireTime(data.accessTokenExpireTime);
-  }).catch(err=>{
+    return data;
+  }catch(err){
     console.log(err.response);
     localStorage.removeItem("refreshToken");
     token.clear();
-    return -1;
-  })
-
-  return 1;
+    return null;
+  }
 }
 
 export const authGet = async (url,options)=>{
-  if(token.check()){//access토큰을 못 쓰는 경우
-    if(reissueAccessToken()=== -1) return null;
-    return await authGet(url,options);
-  }else{//토큰을 쓸 수 있는 경우
+  const refreshToken = localStorage.getItem("refreshToken");
+  
+  if(!refreshToken || refreshToken===""){
+    return null;
+  }
+
+  if(token.check()){//엑세스 토큰이 없거나 사용불가능한 경우
+    const res = await reissueAccessToken();
+    if(res === null) return null;
+
+    const {data} = await axios.get(url,{
+      ...options,
+      headers:{"Authorization" : `Bearer ${token.getAccessToken()}`},
+    });
+
+    if(data.code === 200){
+      return data.data;
+    }
+  }else{//엑세스 토큰이 사용가능한 경우
     try{
       const {data} = await axios.get(url,{
         ...options,
-        headers:{"Authorization" : "Bearer "+token.getAccessToken()},
+        headers:{"Authorization" : `Bearer ${token.getAccessToken()}`},
       });
-
-      return await data.data;
+  
+      if(data.code === 200){
+        return data.data;
+      }
     }catch(err){
-      console.log(err);
+      console.log(err.response);
       return null;
     }
   }
 }
 
 export const authPost = async (url,options)=>{
-  if(token.check()){//access토큰을 못 쓰는 경우
-    if(reissueAccessToken()=== -1) return null;
-    return await authPost(url,options);
-  }else{//토큰을 쓸 수 있는 경우
+  const refreshToken = localStorage.getItem("refreshToken");
+  
+  if(!refreshToken || refreshToken===""){
+    return null;
+  }
+
+  if(token.check()){//엑세스 토큰이 없거나 사용불가능한 경우
+    const res = await reissueAccessToken();
+    if(res === null) return null;
+
+    const {data} = await axios.post(url,{
+      ...options,
+      headers:{"Authorization" : `Bearer ${token.getAccessToken()}`},
+    });
+
+    if(data.code === 200){
+      return data.data;
+    }
+  }else{//엑세스 토큰이 사용가능한 경우
     try{
       const {data} = await axios.post(url,{
         ...options,
@@ -76,16 +106,31 @@ export const authPost = async (url,options)=>{
 }
 
 export const authDelete = async (url,options)=>{
-  if(token.check()){//access토큰을 못 쓰는 경우
-    if(reissueAccessToken()=== -1) return null;
-    return await authDelete(url,options);
-  }else{//토큰을 쓸 수 있는 경우
+  const refreshToken = localStorage.getItem("refreshToken");
+  
+  if(!refreshToken || refreshToken===""){
+    return null;
+  }
+
+  if(token.check()){//엑세스 토큰이 없거나 사용불가능한 경우
+    const res = await reissueAccessToken();
+    if(res === null) return null;
+
+    const {data} = await axios.delete(url,{
+      ...options,
+      headers:{"Authorization" : `Bearer ${token.getAccessToken()}`},
+    });
+
+    if(data.code === 200){
+      return data.data;
+    }
+  }else{//엑세스 토큰이 사용가능한 경우
     try{
       const {data} = await axios.delete(url,{
         ...options,
         headers:{"Authorization" : `Bearer ${token.getAccessToken()}`},
       });
-
+  
       if(data.code === 200){
         return data.data;
       }
@@ -97,10 +142,25 @@ export const authDelete = async (url,options)=>{
 }
 
 export const authPut = async (url,options)=>{
-  if(token.check()){//access토큰을 못 쓰는 경우
-    if(reissueAccessToken()=== -1) return null;
-    return await authPut(url,options);
-  }else{//토큰을 쓸 수 있는 경우
+  const refreshToken = localStorage.getItem("refreshToken");
+  
+  if(!refreshToken || refreshToken===""){
+    return null;
+  }
+
+  if(token.check()){//엑세스 토큰이 없거나 사용불가능한 경우
+    const res = await reissueAccessToken();
+    if(res === null) return null;
+
+    const {data} = await axios.put(url,{
+      ...options,
+      headers:{"Authorization" : `Bearer ${token.getAccessToken()}`},
+    });
+
+    if(data.code === 200){
+      return data.data;
+    }
+  }else{//엑세스 토큰이 사용가능한 경우
     try{
       const {data} = await axios.put(url,{
         ...options,
